Improves the latest transformer caching, resolving some side effects (such as thumbnails).

+++ code/the-fascinator2/trunk/plugins/transformer/ims/src/main/java/au/edu/usq/fascinator/transformer/ims/ImsTransformer.java
@@ -38,6 +38,7 @@ import au.edu.usq.fascinator.common.JsonConfigHelper;
 import au.edu.usq.fascinator.common.MimeTypeUtil;
 import au.edu.usq.fascinator.common.storage.StorageUtils;
 import java.io.FileOutputStream;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
 
 /**
@@ -77,19 +78,25 @@ public class ImsTransformer implements Transformer {
     @Override
     public DigitalObject transform(DigitalObject in)
             throws TransformerException {
+        String sourceId = in.getSourceId();
+        String ext = FilenameUtils.getExtension(sourceId);
+
+        // We are only interested in ZIP files
+        if (!ext.equalsIgnoreCase("zip")) {
+            return in;
+        }
+
         File inFile;
         String filePath = config.get("sourceFile");
         if (filePath != null) {
             inFile = new File(filePath);
         } else {
             try {
-                inFile = File.createTempFile("aperture", ".object");
+                inFile = File.createTempFile("ims", ".zip");
                 inFile.deleteOnExit();
                 FileOutputStream tempFileOut = new FileOutputStream(inFile);
                 // Payload from storage
-                String tempPid = in.getSourceId();
-                Payload tempPayload;
-                tempPayload = in.getPayload(tempPid);
+                Payload tempPayload = in.getPayload(sourceId);
                 // Copy and close
                 IOUtils.copy(tempPayload.open(), tempFileOut);
                 tempPayload.close();
@@ -133,24 +140,22 @@ public class ImsTransformer implements Transformer {
      */
     public DigitalObject createImsPayload(DigitalObject object, File file)
             throws StorageException, IOException {
-        if (file.getAbsolutePath().endsWith(".zip")) {
-            ZipFile zipFile = new ZipFile(file);
-            ZipEntry manifestEntry = zipFile.getEntry(manifestFile);
-            if (manifestEntry != null) {
-                Enumeration<? extends ZipEntry> entries = zipFile.entries();
-                while (entries.hasMoreElements()) {
-                    ZipEntry entry = entries.nextElement();
-                    if (!entry.isDirectory()) {
-                        String name = entry.getName();
-                        Payload imsPayload = StorageUtils
-                                .createOrUpdatePayload(object, name, zipFile
-                                        .getInputStream(entry));
-                        // Set to enrichment
-                        imsPayload.setType(PayloadType.Enrichment);
-                        imsPayload.setLabel(name);
-                        imsPayload.setContentType(MimeTypeUtil
-                                .getMimeType(name));
-                    }
+        ZipFile zipFile = new ZipFile(file);
+        ZipEntry manifestEntry = zipFile.getEntry(manifestFile);
+        if (manifestEntry != null) {
+            Enumeration<? extends ZipEntry> entries = zipFile.entries();
+            while (entries.hasMoreElements()) {
+                ZipEntry entry = entries.nextElement();
+                if (!entry.isDirectory()) {
+                    String name = entry.getName();
+                    Payload imsPayload = StorageUtils
+                            .createOrUpdatePayload(object, name, zipFile
+                                    .getInputStream(entry));
+                    // Set to enrichment
+                    imsPayload.setType(PayloadType.Enrichment);
+                    imsPayload.setLabel(name);
+                    imsPayload.setContentType(MimeTypeUtil
+                            .getMimeType(name));
                 }
             }
         }
